refactor: update board list component
