--- src/components/BoardList/index.module.scss
@@ -1,10 +1,0 @@
-.link {
-  align-self: flex-end;
-  height: 24px;
-  color: rgba(0, 0, 0, 0.54);
-  padding: 8px;
-  &:hover {
-    background-color: rgba(0, 0, 0, 0.04);
-    border-radius: 50%;
-  }
-}
